requirements/base.txt
Django==1.6.1
amqp==1.3.3
anyjson==0.3.3
billiard==3.3.0.13
celery==3.1.6
django-cors-headers==0.12
django-model-utils==1.5.0
django-registration==1.0
djangorestframework==2.3.10
kombu==3.0.8
parse==1.6.3
pytz==2013.8
wsgiref==0.1.2
South
psycopg2
gunicorn
